refactor: 파일 분리
- creator 파일로 돔 요소 생성 로직 분리
- selector 파일로 돔 요소 선택 로직 분리
