+++ src/selector.js
@@ -1,0 +1,16 @@
+const select = {
+  purchaseForm: document.querySelector(".purchase-form"),
+  purchaseInput: document.querySelector(".purchase-input"),
+  userLottoContainer: document.querySelector(".user-lotto-container"),
+  purchaseButton: document.querySelector(".purchase-button"),
+  winningInputContainer: document.querySelector(".winning-input-container"),
+  winningModalContainer: document.querySelector(".modal-container"),
+  winningTableContainer: document.querySelector(
+    ".winning-result-table-container"
+  ),
+  winningDialog: document.querySelector(".winning-result-dialog"),
+  closeButton: document.querySelector(".winning-result-close-button"),
+  restartButton: document.querySelector(".winning-result-restart-button"),
+};
+
+export default select;
